change log Z to M/H
it was pointed out that the Castelli and Kurucz models are not in Log Z but rather metal abundaces [ M/H ]. After checking the papers for this Atlas (and the readme file) I found that indeed it is incorrect. The documentation found for basis models and parameters used to construct the Kurucz 93 models, indicates the same appliees to Kurucz 93 Atlas. Changing the Pysynphot documentation.